Updated openstack/openstack Project: openstack/nova  dc06cf40c44eabcd99ad97486914515a839d2fc4
Use kwargs from compute v4 proxy change_instance_metadata

The args were passed to the compute manager method in the wrong order.
We noticed this in the gate with KeyError: 'uuid' in the logs because of
the LOG.debug statement in change_instance_metadata. Just use kwargs
like rpcapi would normally.

There isn't a unit test for this since the v4 proxy code goes away in
liberty, this is for getting it into stable/kilo.

Closes-Bug: #1444728

Change-Id: Ic988f48d99e626ee5773c97904e09dbf00c5414a