Updated openstack/openstack Project: openstack-infra/project-config  435e5466af9dd8bb0069545dc0041a6b0ce53cfb
Temporarily ignore requirements on master libs

We need to disable checks of master branch requirements in libraries
so we can release updated versions of all libs that have no capped
requirements. We can restore the checks when that is done, and then
we'll look into ways to update the tools to not update caps on
library projects.

This is a continuation of I4e4ac0580b9a1a98cc5cc9ba62d6977324a15296
adding two more libraries to the set.

Change-Id: I55072b64d3d7f85b134fde035cb0ed1b36def90c