Updated openstack/openstack Project: openstack/swift-bench  ae51e6b26fcddc97e3787feccba69db29fa728aa
Fix bug when using --log-level option

Fixes a bug when using the --log-level/-o option. These are currently not
usable, because a condition always exits if there are not exactly two
command-line arguments used (the ip and port).

This patch also ensures that the used log-level settings are valid choices.

Change-Id: I5d64e57b59947c059924fce3e458cfe9fa284318