Updated openstack/openstack Project: openstack/neutron  373d2494b5b2cff5c09d3f7b51b7859c9566baab
Correct typo for matching non-dict ovsdb rows

As can be seen just above, the correct operator for the equality
test is '=' and not '=='. This match isn't currently being used
in the neutron code, but will be used by the OVN driver.

The previous code would also raise NotImplemented when there was
no match.

Change-Id: I17ac85d1ad68d3e207225db300f65c0df1f6e1ad