Updated openstack/openstack Project: openstack/horizon  f5fa5398de3021491aa0d1ba01b20f9e240801ce
Added more details to items per page help text

"Items per page" value setting is applicable only to those pages 
that have API supported pagination.  For example, if "Items per page"
value is set to say "3" and if we go to Admin > Identity Panel > Users
page, more than "3" items will be displayed if present.  This is
confusing to users.  So as part of this commit, the help text is modified
to reflect this limitation.

Change-Id: Ibbd4771829f0c312134eb275c797780be2b78410
Closes-Bug: 1431364