Updated openstack/openstack Project: openstack/horizon  322530a5a50e6ccceb112317d1bda1626a45af22
removed asterisk for admin state in network create

In network create and update, the admin_state parameter is optional
but in horizon forms, asterisk is shown for admin state choice field which is wrong.

In this patch, asterisk mark is removed for admin state choice field in network create and update forms.

Change-Id: I349915f8fbe496cf79d33cd54e9c405799eb61e2
Closes-Bug: #1441511