Updated openstack/openstack Project: openstack/nova  e507a0515a88efe10c92b24965251e831c45327d
tests: introduce a NUMAServersTest class

Introducing a new NUMAServersTest class and two functional tests
to validate a correct creation of instances with a requested NUMA
topology, as well as to validate that instance won't be created
when the requested topology can not be satisfied.

Also, adding a method automatically generate the numa
topology xml, using the libvirt config classes.

Change-Id: Ibdbce1c54003ae708b6c7f4b0874ca1ad96bfa07