Updated openstack/openstack Project: openstack/requirements  b0ced14db71615bd59a2f22ca0585807c7289654
Bump PyECLib version from 1.0.3 to 1.0.7

PyECLib is a requirement for Erasure Coding in Swift (a critical
new feature being introduced in Kilo).  In addition to fixing
several bugs, 1.0.7 eliminates the need for a few work-around code
in Swift.  This code was only to hide issues in the current version,
but it also ends up breaking some third-party integration.  In order
to enable expected functionality and to avoid dealing with deprecation
issues right from the beginning, we hope to bump the minium PyECLib
requirement to 1.0.7.

Change-Id: I03e059e7335656c22be28ffd6157b56e13bdfc1b