Updated openstack/openstack Project: openstack/openstackdocstheme  415893ae82e446a7d3850c24d06baec62b454c7b
Openstackdocstheme: log-a-bug prefills wrong source file

When clicking on the bug icon on RST-based documentation
pages, the launchpad page that is opened sports a wrong
link to the source file.
Current code assumes that all source files reside in the
same directory as the conf.py file. This assumption is
wrong; source files may well live in subdirectories.

The fix is in script_footer.html. The Javascript code
used to use Jinja2 variable "sourcename", which doesn't
contain the subdirectory of a source file.
This is changed to "pagename", which does.

Change-Id: I3289f6c7a672fbc20c0d9e1183c65ffa49d4db9d
Closes-Bug: #1446444