Updated openstack/openstack Project: openstack/diskimage-builder  e9ac9c98239666e558dba38c608d045a6a12cf1d
Open MKFS_OPTS for extension in disk-image-create

At present, MKFS_OPTS is closed for modification. The ability
to extend the set of MKFS_OPTS adds a great deal of power for
knowledgeable end-users. (And in some specific circumstances,
it is vital to success, as in the case of building RHEL/CentOS
6 images from RHEL/CentOS 7 hosts, in which case -O ^64bit is
required in order for the image to boot.)

Change-Id: I714e86a5a413779e63f598fbbb5a79d23cf6d8c3