Updated openstack/openstack Project: openstack/horizon  03f60545981a06dbe590ef2d30642a5ff0ddaa5b
Fix unnecessary dotted focus outline in tab bar

In Firefox, the tab bar in "Volumes" and "System Information" pages 
displays unnecessary dotted focus outline.  This is not the case with
other browsers like Chrome or Safari.  This is fixed by adding
'outline:none' to :focus in the tab bar styling (there is already CSS
for :focus).

Change-Id: I3f7065003d3c0ec20fa1d3f0189e208bb36e9b94
Closes-Bug: 1431341