Updated openstack/openstack Project: openstack-infra/system-config  714ec5197dcaed1dab0c64476db5a23ba5c1fd6b
Add redirects from ci.o.o to docs.o.o/infra

We're moving the content of http://ci.openstack.org to
http://docs.openstack.org/infra and will use this to redirect old
deep links once the content is in place. It's safe to land this in
the meantime since it won't be hit until we switch DNS for
ci.openstack.org.

Change-Id: I6c6091e1dd18931211691972c408db8f3bf81719