Updated openstack/openstack Project: openstack-infra/project-config  a96d097850bef06d683bd8cdf20e428ff17bc630
networking-ovn: enable mysql and rabbitmq

The devstack job for the networking-ovn repo overrides the list of
enabled services to be a minimal set.  Unfortunately, it's missing
mysql and rabbitmq which are required for devstack to start up
successfully with the other services that are enabled.

This patch also fixes enabling keystone by specifying it as 'key'
instead of 'keystone'.

Change-Id: I29a3b2c03559aeaba7c08431bfbed1fe21762666