Updated openstack/openstack Project: openstack/nova  a911cadfdc6b5ffa9b13dda00b7f8368fe67b17a
VMware: use oslo.vmware methods for handling tokens

When the VC has paginated results it returns a token. We now
make use of utilities provided by oslo.vmware to continue the
retrieval of the tokens or to cancel the retrieval of the
tokens.

This makes use of methods provided by oslo.vmware. It is far
cleaner and robust than the previous code.

Change-Id: I389496418d9b97a32561c5c2183ea549bfa908b2