Updated openstack/openstack Project: openstack-infra/project-config  5787cae3c906217c322bbe35b2938e371936ab75
Ensure we run the full tempest job serially without admin

This commit fixes a previous oversight to the non-admin tempest job
definition. Previously the job was configured to run the smoke tagged
subset of tempest in parallel. However we need to run the full suite
serially with the non-admin job. This adds the missing flag to the
definition.

Change-Id: Ie54d282d222b79dcc36fd3087665888f5094bf36