Updated openstack/openstack Project: openstack/nova  07593a14049b1990ab9ab6aa078bb96b2905e232
Validate server group affinity policy

In git commit a79ecbe Russell Bryant submitted a partial fix for a race
condition when booting an instance as part of a server group with an
"anti-affinity" scheduler policy.

It's possible to hit a similar race condition for server groups with
the "affinity" policy. Suppose we create a new group and then create two
instances simultaneously. The scheduler sees an empty group for each,
assigns them to different compute nodes, and the policy is violated.

To guard against this, we extend _validate_instance_group_policy()
to cover the "affinity" case as well as "anti-affinity".

Partial-Bug: #1423648
Change-Id: Icf95390a128e2062293e1f5b7b78fe79747f5f27