Updated openstack/openstack Project: openstack/manila  85f067cff559366797a075672baa03e26caffd16
Add mount/umount in scenario tests

This adds mounting and unmounting of a created share. In order
to be sure that the nfs/cfis tools are available in the used image
it's using the default Ubuntu nfs/cifs image. Since cloudinit
is not supported for this image it's using username/password
authentication as workaround.

Change-Id: I77720ba5beb836614728081bbf74b44991d6e284
Partially-implements: blueprint scenario-tests