Updated openstack/openstack Project: openstack/nova  617f28d6317ff653317e24f64603a49d128104c9
Add field name to error messages in object type checking

The coerce() methods in object fields generate ValueError
exceptions if an incorrect type is used. Most of the messages
do not indicate which field was involved although that information
is available in the coerce() methods. This change adds the
field name to the error messages.

Change-Id: Ic94d52f6019c5a5a291675e321f96f5c65beaf10
Closes-Bug:  #1430681