Updated openstack/openstack Project: openstack/diskimage-builder  e61ffb7fc84bb072b65bacbddad0d482fc36b1bc
Don't trace RHEL Registration scripts

We don't want to trace the RHEL registration scripts because that
is likely to log things like passwords and activation keys.  To
still allow for debugging failed runs, add sanitized logging of
the arguments passed to the registration commands, since that is
the part of the process where problems are most likely to manifest.

Change-Id: I0f661e9c152f43b814fda61211bd56ba93e3b9dc