Updated openstack/openstack Project: openstack-infra/project-config  9084b49579f850fb3167eb64bb8a2965f9b640ce
Use Oslo core teams for stable maint, too

The Oslo team has decided to fold stable maintenance work into the
responsibilities of the core teams (we're small and the overlap on the
teams was approaching 100% anyway).

This change removes the exclusiveGroupPermissions settings for the stable
branches, making the remaining settings additive. It then adds the
stable-maint-core team so they can assist the various Oslo teams.

Change-Id: Ie4031cbe1ebe7f089a70e1946e234dac07727e85