Updated openstack/openstack Project: openstack-infra/project-config  323e2d646e74654ac5322a904d08b4cea6ac202a
The sahara-tox-py27-scenario-unit job should be voting since kilo

This jobs already working ok for some some time as non-voting check job
and now we're ready to enable it as voting for branches starts with kilo.
In addition to it, this change is needed to stop running this job for
stable/juno where this tests were not yet implemented.

Change-Id: I6f8b11d03dc944b4b3595750bb0b090fad74ee61