Updated openstack/openstack Project: openstack-infra/shade  1f21b4637208e6edc7facff08792daac5a63a2e7
Add test for user_cache

Since we intend to invalidate the user_cache, and properties and caching
decorators don't play nicely, we split this into the property and a
getter which can be called explicitly.

Note that this test also uncovered a bug in user_cache, which was that
it called a non-existent task.

Change-Id: Ica04a82242cf4724155f99de5b8d3d9c356a8ec4