Updated openstack/openstack Project: openstack/governance  361a2dcbd5250020e0fcb9d892b816e81d9242b1
Let projects add repos without prior-approval

Some OpenStack projects are adding git repos at a moderate pace
and generally the TC is rubber-stamping them as long as the PTL
approves.  Explicitly codify that this is desirable -- the TC
doesn't really care how many git repos it takes to accomplish a
task, so as long as a new repo fits within the existing mission
statement, indicate that it is permissible to add it without
prior TC approval.

This is not strictly a "new project requirement", however, this
document seems to be the best existing place to add it as it
relates to project creation.

Change-Id: I34acfe4ffe5a0af6a941e70a5315a07d036e6c24