Updated openstack/openstack Project: openstack-infra/system-config  d4e3158ef9c0f6281618c8f2e5608b64b339b0e8
Do not install linux-headers-virtual for trusty or later

In trusty and later, linux-headers-virtual is now a transitional package
that simply depends on linux-headers-generic.  This package does not
exist for the aarch64/arm64 architecture and causes installation to fail.

This change continues to specify both linux-headers-virtual and
linux-headers-generic for precise, but only specifies
linux-headers-generic for all other releases.

Change-Id: I64542d818b22c57c6bf77173f83f148e492683ab