Updated openstack/openstack Project: openstack-infra/subunit2sql  76e0c04e03ea81cea89736b8e15e409a9c62791e
Add command to calculate failure count and graph it

This commit adds a new analysis interface for dealing with status
percentages of a single test. It will print out the percentage
failure, success, and skips of a test over time. It additionally
will graph the failure count per day over time, and optionally
do the same with the success and skip counts.

Change-Id: I485afbad5b89fc702a2d9a739ab3f922fbffd4bc