Updated openstack/openstack Project: openstack/trove  5e4b98a1e17c93bdb6f50bec128862679a118f4c
Moves taskmanager-common-code for clusters

Found that the following methods are quite similar in appearance:
1. update_statuses_on_failure
2. _all_instances_ready
3. get_ip
4. get_guest

Also, it seems that these methods would be needed
for any future cluster implementations,
so it would be better if they have some common residence.

Hence, moved these methods to taskmanager.models.ClusterTasks,
which is parent to both MongoDbClusterTasks and VerticaClusterTasks.

Change-Id: I39f7dc2f5ab318ac095b8bf4742c83ab2ca9a499