Updated openstack/openstack Project: openstack-infra/shade  5b9ff8cba2d3f3eb58b1b62a2db1ad2190ce88a0
Update volume API for new getters and dict retval

This switches the volume API to use the new get/list/search style.
Doing this necessitated that we fix another issue with the volume
API and NOT return the volume object as received from cinder, but
rather convert it to a dict using meta.obj_to_dict.

Change-Id: I81ba97d2b4fc9465e1b3538e97f5ceb9bef4f7e8