Updated openstack/openstack Project: openstack/oslo-incubator  d5edda00b4eca65d57f94bd0ac1b790e6d1f732e
service child process normal SIGTERM exit

service.py had some code where the child process would catch the
SIGTERM from the parent just so it could exit with 1 status rather
than with an indication that it exited due to SIGTERM. When
shutting down the parent doesn't care in what way the child ended,
only that they're all gone, so this code is unnecessary.

Also, for some reason this caused the child to never exit while
there was an open connection from a client. Probably something
with eventlet and signal handling.

Change-Id: Ie51669bd278288b768311ddf56ad31a2f28cc7ab
Closes-Bug: #1446583