Updated openstack/openstack Project: openstack-infra/infra-manual  a5050b373c9492b3b69ee0a247df3fd285f34bea
Project does not mean repository? Inconceivable!

You keep using that word. I do not think it means what
you think it means.

Previously "project" meant a variety of different things
depending on the context in which it was used. Now that
the TC has decreed "project" is a grouping of source
code repositories managed by a "project team" we need to
be extra careful not to conflate these meanings in our
documentation, particularly for the sake of minimally
fluent readers and those not yet familiar with our
community's unique terminologies.

Change-Id: I574ab8e2c94eadce76cff9f7ea561dfa5d28b89c