Updated openstack/openstack Project: openstack/neutron  6661dcda6bf97b1f7b4d61acce051a5b00c4efc7
Added note about removing bridge from mappings

A bridge removed from mappings is not managed by the OVS agent anymore,
but continues to be connected to the integration bridge.
Added a note about it in the config so that deployers make sure they
don't end up in a sticky situation.

Change-Id: I8992f842046651e0f231c6bf08b65efa07056757
Related-bug: #1436267