Updated openstack/openstack Project: openstack/nova  2fd3d7e26e7b486eebb25a1f0f93b0cfca059371
remove downgrade support from our database migrations

The TC approved a cross project spec to stop putting db downgrade into
project migration repos:
https://github.com/openstack/openstack-specs/blob/master/specs/no-downward-sql-migration.rst
This was done after quite a bit of feedback from the operator
community, and everyone was good with it.

This implements that spec for Nova, adjust the tests to pass
accordingly, and changes the test which requires downgrade functions
to one that prohibits them.

Implements bp nova-no-downward-sql-migration

Change-Id: I9b164b748998530d62bbc6baf356a9c3f61584cc