Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  78868c69506171bccf715d9db92bfec67ecb2098
Add cron job to run weekly prune the data set of db

This commit adds a new cron job to the subunit2sql server manifest to
periodically run a data expiration on the subunit2sql db. This will
delete any data in the DB which is over 186 days old which is roughly
6 months. This should enable us to keep 1 releases worth of data in
the DB at a time.

Change-Id: I0e52a1943d75fedfefa8beb2b9f1fcdb406af220