Updated openstack/openstack Project: openstack/python-novaclient  d0263a76a38082f1bae32b2d60f56e8cc7d96fe3
Fix displaying of an unavailable flavor of a showing instance

Since then an instance was booted with a flavor, the flavor could be
deleted or become unavailable (private flavors). As a result novaclient
fails to show the instance with member role user.

This patch set adds handling of unavailable flavors to showing of
instances as it is done for unavailable images.

Closes-Bug: #1366168

Change-Id: I810dcac814b7523313112bb9151754659b85df51