Updated openstack/openstack Project: openstack/cinder  8d27e2939617053eb5c26e9479140578ef257737
Remove force check from copy_volume_to_image

The upload_volume_to_image method allows a force parameter that will
upload a volume even though the volume is attached/in-use. A user can
get away with this with the LVM driver because the LVM backing is
local to the Cinder worker node that's pushing the bits to Glance.

The problem is, this only works for local storage, it won't work
with any iSCSI devices because they can't do multi-attach. Also,
the reason we required that a volume NOT be in-use for this
operation is because we have no way of keeping the guest Instance
from writing to the volume while we're uploading, and corrupting the data.

This has been exposed like this for several releases, so removing it
now likely would not be a good user experience.  Instead, this
patch add a config option to enable/disable it (default is to
disable), and deployers can choose whether they would like to
allow the use of --force True or not.

DocImpact Disables the --force option to copy-volume-to-image and
          introduces "allow_force_upload" boolean option to re-enable

Change-Id: I1210186689dea204601356f8d08805c6cb6f017c
Closes-Bug: 1446954