Updated openstack/openstack Project: openstack/trove-integration  a35963165a4e589cd57f3040f616fefce528e5e6
Added support for OSProfiler in redstack

A flag (ENABLE_PROFILER) was added to redstack.rc that controls
whether OSProfiler is enabled.  Since this depends on Ceilometer,
Ceilometer is also installed if set to true.

Several configuration variables were added as well to redstack.rc
that can be used to control the behaviour of the profiler.

ENABLE_PROFILER is set to 'false' by default, but can be turned
on easily by adding 'ENABLE_PROFILER=true' to options.rc or by
the command line 'ENABLE_PROFILER=true ./redstack install'.

Change-Id: I981809d5c5fd5ac9a7d72713e90d93845fc3a747
Related-Bug: #1421859