Updated openstack/openstack Project: openstack/heat  308143e3d2f6b98420c84fe8cbd8bcf4fe6f9645
Fix ResourceGroup validate when there are removal_policies

We can't call child_template() from a validate in ResourceGroup
as it accesses nested(). So move the validation to validate_nested_stack()
so we can control the test template used for validation.

Unfortunately coverage for this got removed during recent test rework
and it looks like the interface has got broken.

Co-Authored-by: Angus Salkeld <asalkeld@mirantis.com>
Change-Id: Ibea738a343847736b041cc49a2c486fa71e562d9
Closes-Bug: #1446861