Updated openstack/openstack Project: openstack/ironic  fee6b2bcfb723dba2f128114177f6344e34fc098
Remove ironic compute driver and sched manager

This removes the deprecated ironic compute driver and scheduler
manager. The compute driver and scheduler manager have been in
nova's tree since Juno and this was to have been deprecated in
the Kilo release. It is now Liberty, so it is about time :)

DocImpact

Change-Id: I8c3c35ca5a9a163e689a4c856b1efd557779b9fb