Updated openstack/openstack Project: openstack/python-novaclient  3251bc430fbe406b8aabbea56418c1bd20b2ed54
Handle binary userdata files such as gzip

The current code in the servers api assumes only text files will be
provided as userdata. That is, the data is utf-8 encoded before it
is base64 encoded and sent. This breaks for binary userdata files,
(example, gzip files) as they can't be utf-8 encoded.

This change ignores specifically the exceptions that are raised
when utf-8 encoding is attempted on non-text data: AttributeError
and UnicodeDecodeError. These exceptions will be ignored to let
binary files proceed to the base64 encoding step.

Closes-Bug: #1419859

Change-Id: Ie96283f6ff892ae30485722cf7c3ec40f3f1b5df