Updated openstack/openstack Project: openstack/nova  d53ef7f0aa94e7f84f94f76b1514526d6932781d
libvirt: Allow discrete online pCPUs for pinning

Currently, the check for cpu pinning is a little broad and
it doesn't allow pinning for discrete online pcpu sets if the
cpuset ID exceeds the total number of online pcpus.

The fix is for NUMA topology enabled hosts, where the list of
online pcpus are obtained and compared with the list of IDs which
are requested as a part of cpu pinning. If the IDs form a subset
of the total online pcpus on the host, pinning is allowed.
If not, old behavior is retained.

Change-Id: I8cbd938f99e1eae9eeaa3bb27f492c95734bbfab
Closes-Bug: #1442546