Updated openstack/openstack Project: openstack/nova  fcb77e7678e887e8a83211ef530bc531068bbdf6
Libvirt: Fix error message when unable to preallocate image

For this error message,

LOG.error(_LE('Unable to preallocate_images=%(imgs)s at path: '
              '%(path)s'), {'imgs': CONF.preallocate_images,
                            'path': self.path})

CONF.preallocate_images can only be set to 'none' and 'space'
so if error message come out, it would be

Unable to preallocate_images=none(or space) at path: /path/to/image.
which is not correct.

Change-Id: Id122a5e1ad1d0441d6b75d081b2be3f276439a2a