Updated openstack/openstack Project: openstack/requirements  372a19f379422a49127b756b759964683811d7d1
Remove OpenStack library caps for liberty

We now have a stable/kilo branch of the requirements repo, so we can
remove the caps on Oslo libraries for the liberty cycle. This change
also uncaps libraries that were managed by Oslo when the caps were
introduced, but are part of other projects now.


Change-Id: Ica92602074ef5848c6b1b61196087848b055f4f9