Updated openstack/openstack Project: openstack/tempest  8912814caa1564247ae01c3c3fd518d1f4a012ef
Switch all uses of urllib and HTMLParser to import from six

As part of enabling python3 support in tempest we need to be able to
handle urllib, HTMLParser, and urllib2 usage in both python2 and
python3. Six provides a compat layer for doing this, so this commit
moves all uses of urllib2, HTMLParser, and urllib to get it through
six.

Change-Id: I81e348ac79001ac94bdb92a9b60c7dca4c93686a