Updated openstack/openstack Project: openstack/tempest  1c517a22cadf916312fb302e5b1c369df398874a
Switch use of hashing strings to unicode in accounts

This commit switches all the uses of a str cast on hashed indexes in
the accounts provider code to use six.string_type casts instead. This
is to ensure python3 compat which uses unicode as the default string
type. However, as part of this the strings need to be encoded
manually before they can be consumed by hashlib. As part of this
the unit tests also needed to be updated slightly to take this into
account.

Change-Id: Ic77997fbed9ff1da4cde7f35955d0381a13c49bf