Updated openstack/openstack Project: openstack/nova  c4632842645c3b58c03d4c94b71646721404ab7a
Override update_available_resources interval

Add a config option 'update_resources_interval' to be able to set how
often the update_available_resources periodic task should run.

When a single nova-compute is managing a lot of nodes, we need the
ability to tune the frequency of this task.

The default value of 0 means to run on every periodic task interval
(same as current/old behavior). Setting it to a negative value will
prevent it from running at all. Setting it to a positive value will
determine the number of seconds between calls.

DocImpact
This adds a new CONF option 'update_resources_interval' that controls
how often the update_available_resource() method is called.

Co-Authored-By: Ed Leafe <ed@leafe.com>

Change-Id: I8986d8d97dbf348e219c5f1f0a62092115c096d5