Updated openstack/openstack Project: openstack/nova  93de6e1d331fa7e7369e2047b2d2f1459fa1cd34
change the reboot rpc call to local reboot

Instances fail due to the reboot operation using the rpc call,
because the rpc call can fail due to the network impact.
Due to  the rebooted instances are just at the current host,
There is no need using the rpc call to reboot the instance.

Change-Id: I3d435cf9ca5f722ed3fc4325e7331f29a73198d5
Closes-bug:#1439493