Updated openstack/openstack Project: openstack/trove  0ba5aeea2dad870db37692531e7c53ca07209f51
Add unix_socket_directories setting for pgsql

When an existing pgsql instance is resized, we restart in a state
where /var/run/postgresql does not exist and the postgresql
service is initially disabled. If /var/run/postgresql doesn't exist
and unix_socket_directories is not specified (as was the case for
us), the ubuntu/debian scripts that wrap many postgresql binaries
including psql will fail

A bug 1446811 has been filed upstream to the maintainers of
package postgresql-common.

Change-Id: I69a8f13f3577b3c37a75d7323eaabdf396b5b5e7
Closes-Bug: #1445651