Updated openstack/openstack Project: openstack-dev/devstack  5ea9fb13056f9aa738711acf445b78914b734570
Fix issue with ml2 plugin on using provider network

This commit fixes the issue that ml2 plugin
sets 'flat_networks' in ml2 configuration file as
empty.  The value of 'flat_networks' need to be set 
as the name of the physical network that was specified 
in the localrc file (or it's default value).

Change-Id: Ib4c31f6576da57534b36aefebd1ca8cd397c6c1a