Updated openstack/openstack Project: openstack/neutron  78b32aad9031bfbf0dd478a7522b6075e610d98e
Log RPC initialization in L3 service plugin and ML2

Under certain conditions the messaging server may be up but
not responding. In this case the Neutron server will fail to start
silently, making it pretty hard to track down the issue without
looking through a lot of code and adding a bunch of random logging.

Change-Id: I6a562476f2789386a020db7b21b9349c4c58c30c